86cyrzxwu: Moved everything to do with the profile to being in its own separate directory under ui-models, and refactored all imports to the new directory
