+++ app/client/store.ts
@@ -4,7 +4,7 @@ import { agentDashboardSlice } from "./ui-modules/role-dashboard/agent-dashboard
 import { landlordDashboardSlice } from "./ui-modules/role-dashboard/landlord-dashboard/state/landlord-dashboard-slice";
 import { guestLandingPageSlice } from "./ui-modules/guest-landing-page/state/reducers/guest-landing-page-slice";
 import { settingsPageSlice } from "./ui-modules/settings-page/state/reducers/settings-page-slice";
-import { profileSlice } from "./ui-modules/role-dashboard/agent-dashboard/state/profile-slice";
+import { profileSlice } from "./ui-modules/profiles/state/profile-slice";
 import { loginFormSlice } from "./ui-modules/user-authentication/state/reducers/login-form-slice";
 import { signupFormSlice } from "./ui-modules/user-authentication/state/reducers/signup-form-slice";
 import { TypedUseSelectorHook, useDispatch, useSelector } from "react-redux";
