+++ app/client/ui-modules/profiles/state/profile-slice.ts
@@ -1,5 +1,5 @@
 import { createSlice, PayloadAction } from "@reduxjs/toolkit";
-import { RootState } from "../../../../store";
+import { RootState } from "../../../store";
 
 interface ProfileData {
   firstName: string;
